[FSSDK-12296] review update
